Model fixes
Properly handle air model and items with no item_model component (should render empty slot).
I figured out why banners and pots had different lighting, but I decided to keep it.
Changed the sorting of the block sheet a little. I would prefer for items and models to be interspersed, but it's not really appropriate for get_model_sprites to take an item model name parameter.
Lastly, I realized that blocks with tints and arbitrary faces with tintindex simply cannot be rendered dynamically. In vanilla, this is only the grass block. Luckily, its tint is hardcoded, so we can just render it as an item, the only pre-tinted item in the block grid. Leaves could do this too, but since all of their faces are tinted, we can still do it with text component color, which I prefer. You might think I could split the grass block into layers like with flat sprites. That would involve partitioning the model into smaller fake models of each tint index, then drawing them in order. The problem with that is that we don't know in general which elements logically cover each other.