Work on porting 1.14 to beet
The script doesn't have to be good (and it isn't), but we should be able to generate the pack without relying on long-since missing files on my computer.
All the subsequent versions will be ported after this one, and should be easier.
Port is not quite complete. The block sheet currently has uncolored banners, an untinted grass block, and a weirdly positioned dragon head. I've already added manual fixes for chests and other weird things that can't be rendered straightforwardly in a version this old.
As a plus of this porting, I think the old block sheet used the wrong texture for spruce leaves, and some spawn eggs and blocks like diorite were missing pixels. Stained glass also just used the gray inventory background. All of those should be fixed when this is done.