Pull out plugins part 1
I thought I was almost done with this port, but I've become unhappy with the three copy/pasted hardcoded plugins.
The new plan is to pull out all the python code to the top level. Just one file for now, but I could see having more. I'm going to merge the individual python plugins into this, while preserving their minute differences.
I also plan to reduce hardcoding as much as possible. For example, the block and item lists in the pre-1.20.5 scripts (the ones that currently have a giant "blocknames" list) are now deduced from the pack in a similar way to the later scripts.
Spawn egg colors, rather than being hardcoded, are sourced from 1.21.4, when they were visible in the jar file. Only specific changes need to be hardcoded. Item durability can also be sourced from versions later than the one we're building.
The version metadata is specified much better. You just write the target version (or two if a range), and the pack formats and data version are fetched automatically.
This is still a work-in-progress. I need to make sure everything builds the same way it did before.
Up next, we need to port the assembly of the item/block sheets for pre-1.16.