Use bolt for functions
Bolt only has builtin command trees for 1.16 and above, so we need to provide our own.
Bolt "modules" are weird, and go in the datapack folder for some reason. They're ending up in the output now, and will need to be removed.
Comments are stripped out for now, because the alternative is awkward spacing in generated functions.
Overrides are implemented correctly in a non-hardcoded way. Overlays, including both textures and stack/durability, are not yet ported.