Return to vanilla model resolver
Now that the vanilla getter is lazy, we can stick our own vanilla into its private field before it resolves. This is not great, but it's better than forking.
Now that add_model_dict_task exists, we can render our fake models without writing them to the pack. This is nice.
Now that tasks with no path save to the saved_img field, we can grab them without writing them to the output pack. I still don't love the task model, but this is still better than before. Since the tasks transform while resolving, we can't really keep track of them beyond their order in the task list, so we just zip with the item keys to find their matching tasks. This is iffy but works.