Banner pattern overlays
Prior to hex color text components in 1.16, we need separate characters for each dye color.
This was also the first usage of block overlays in the v1 plugin, so I encapsulated the grid building and rendering to a class that keeps track of whether sprites are overlays or not, so we can use that information when generating the translations.
The banner pattern textures use black instead of transparency in 1.14, so we cheat by looking at textures from a later version.