Dither transparency of gradient banners and stained glass
Doesn't work for flat items unfortunately, so stained glass panes look solid
An alternative fix would be to put the inventory gray background behind the block grid (and include pane textures in the pack with the same background), but it's more important to me to keep compatibility with the user's texture packs as much as possible. Since the tooltip background comes from the user's pack, we can't make any assumptions about it.